+++ CHANGELOG.md
@@ -10,7 +10,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - LACLI001 - 13/04/2023 - Add the RoundPanel component.
 - LACLI001 - 14/04/2023 - Add the ChangeColorFactory.
 - LACLI001 - 14/04/2023 - Add the MenuLabel.
-- LACLI001 - 14/04/2023 - Add the CircleGraphic.
+- LACLI001 - 15/04/2023 - Add the CircleGraphic.
+- LACLI001 - 15/04/2023 - Add the CircleGraphicRAM.
 
 
 ### Changed